+++ run-csaf.sh
@@ -1,9 +1,11 @@
 #!/bin/bash
 JUPYTER=0
+JUPYTER_DIR=docs/notebooks
 source .common.sh
 
 validate_dir
 
+EXAMPLES="{f16-shield, f16-simple, f16-llc-analyze, f16-llc-nn, inv-pendulum}"
 print_help() {
     printf "\033[1mCSAF\033[0m\n"
     printf "    Control System Analysis Framework (CSAF) is a middleware framework that
@@ -11,15 +13,16 @@ print_help() {
     loop topologies and component implementations are specified independently of
     the middleware.\n"
     printf "\n\033[1mUSAGE\033[0m\n"
-    printf "   -e      the name of the example { f16-shield, f16-simple, f16-llc, inv-pendulum }\n"
-    printf "   -c      the name of the model config file (must be in the same directory as your system)\n"
-    printf "   -d      fully qualified path to the directory defining the model system\n"
+    printf "   -e      the name of the example ${EXAMPLES}\n"
+    printf "   -c      the name of the config file\n"
+    printf "   -d      fully qualified path to the directory defining the control system\n"
     printf "   -f      name of the job config file (must be in the same directory as your system)\n"
+    printf "   -h      prints the help menu\n"
     printf "   -j      launch a jupyter notebook\n"
     printf "   -l      build the image locally\n"
     printf "   -n      run CSAF natively\n"
+    printf "   -p      specify root jupyter notebook directory\n"
     printf "   -t      the tag of the image { stable, edge, latest }\n"
-    printf "   -h      prints the help menu\n"
     printf "   -x      clear the output for a particular example/config"
     printf "\n\n"
     printf "\033[1mEXAMPLES\033[0m\n"
@@ -37,9 +40,11 @@ print_help() {
     printf "    ./run-csaf.sh -e f16-shield -f f16_job_conf.toml\n"
     printf "Clear generated outputs for f16 example:\n"
     printf "    ./run-csaf.sh -e f16-simple -x\n"
+    printf "Start jupyter notebooks from this (CSAF root) directory (instead of ./docs/notebooks):\n"
+    printf "    ./run-csaf.sh -e f16-llc-nn -j -p .\n"
 }
 
-while getopts ":c:d:e:f:t:jlhnx:" opt; do
+while getopts "c:d:e:f:hjlnp:t:x" opt; do
     case ${opt} in
         c )
             CONFIG_NAME=`basename $OPTARG`
@@ -61,6 +66,9 @@ while getopts ":c:d:e:f:t:jlhnx:" opt; do
             ;;
         n )
             NATIVE=1
+            ;;
+        p )
+            JUPYTER_DIR=$OPTARG
             ;;
         t )
             IMAGE_TAG=$OPTARG
@@ -98,12 +106,16 @@ then
             CONFIG_NAME="inv_pendulum_config.toml"
             CSAF_LOC=${PWD}/"examples/inverted-pendulum"
             ;;
-	"f16-llc")
+        "f16-llc-nn")
+            CONFIG_NAME="f16_llc_nn_config.toml"
+            CSAF_LOC=${PWD}/"examples/f16"
+            ;;
+        "f16-llc-analyze")
             CONFIG_NAME="f16_llc_analyze_config.toml"
             CSAF_LOC=${PWD}/"examples/f16"
             ;;
         *)
-            show_error_and_exit "Unknown example: ${EXAMPLE_NAME} Please use one of [f16-shield, f16-simple, inv-pendulum]"
+            show_error_and_exit "Unknown example: ${EXAMPLE_NAME} Please use one of ${EXAMPLES}"
             ;;
     esac
 fi
@@ -160,16 +172,20 @@ case $OS in
     ;;
 esac
 
+if [[ ${JUPYTER} -eq 1 ]] ; then
+    echo "Jupyter notebook root directory is ${JUPYTER_DIR}"
+fi
+
 if [[ ${NATIVE} -eq 1 ]] ; then
     if [[ ${JUPYTER} -eq 1 ]] ; then
-        jupyter notebook --no-browser --notebook-dir=${PWD}/docs/notebooks
+        jupyter notebook --no-browser --notebook-dir=${PWD}/${JUPYTER_DIR}
     else
         python3 "src/run_system.py" ${CSAF_LOC} ${CONFIG_NAME} ${JOB_CONFIG_PATH}
     fi
 else
     if [[ ${JUPYTER} -eq 1 ]] ; then
 	 docker run ${JUPYTER_NETWORK} -it -v ${CSAF_LOC}:/csaf-system \
-            -v ${PWD}/src:/app -v ${PWD}/docs/notebooks:/notebooks \
+            -v ${PWD}/src:/app -v ${PWD}/${JUPYTER_DIR}:/notebooks \
             ${IMAGE_NAME}:${IMAGE_TAG} "jupyter" "notebook" "--port=8888" \
             "--no-browser" "--ip=0.0.0.0" "--allow-root" "--notebook-dir=/notebooks"
     else